requirements-dev.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
cssselect2==0.6.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.6
django-admin-interface==0.19.1
django-colorfield==0.7.2
django-filer==2.2.2
django-filter==22.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==2.8.0
django-js-asset==2.0.0
django-mptt==0.13.4
django-polymorphic==3.1.0
django-smart-selects==1.6.0
djangorestframework==3.13.1
easy-thumbnails==2.8.3
et-xmlfile==1.1.0
importlib-metadata==4.12.0
lxml==4.9.1
Markdown==3.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
pycodestyle==2.9.0
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
reportlab==3.6.11
six==1.16.0
sqlparse==0.4.2
svglib==1.3.0
tablib==3.2.1
tinycss2==1.1.1
toml==0.10.2
tzdata==2022.1
Unidecode==1.1.2
webencodings==0.5.1
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.1
